Refactor android platform: create a DemixingPlugin
